rename file and use proper field type
